Docs: Update Remote Config instructions
The setup instructions in the README are updated to reflect the process of uploading a JSON file for Firebase Remote Config.

Transformed the file remote_config_defaults.xml to json format `remote_config_defaults.json`. This new file is added to establish default settings for the application's remote config feature flags and prompts.